+++ src/adminDashboard/products.java
@@ -336,17 +336,21 @@ public class products extends javax.swing.JFrame {
             return;
         }
 
-        DefaultTableModel model = (DefaultTableModel) items.getModel();
+        int confirm = JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this product? This will remove it from the inventory.", "Confirm Product Deletion", JOptionPane.YES_NO_OPTION,  JOptionPane.WARNING_MESSAGE);
 
-        String id = model.getValueAt(selectedRow, 0).toString();
+        if (confirm == JOptionPane.YES_OPTION) {
+            DefaultTableModel model = (DefaultTableModel) items.getModel();
+            String id = model.getValueAt(selectedRow, 0).toString();
 
-        config conf = new config();
-        String sql = "DELETE FROM tbl_items WHERE i_id = ?";
-        conf.deleteRecord(sql, id);
+            config conf = new config();
+            String sql = "DELETE FROM tbl_items WHERE i_id = ?";
 
-        model.removeRow(selectedRow);
+            conf.deleteRecord(sql, id);
 
-        JOptionPane.showMessageDialog(null, "Deleted Successfully!");
+            model.removeRow(selectedRow);
+
+            JOptionPane.showMessageDialog(null, "Product Deleted Successfully!");
+        }
     }//GEN-LAST:event_jButton5ActionPerformed
 
     private void jButton6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton6ActionPerformed
@@ -452,45 +456,7 @@ public class products extends javax.swing.JFrame {
     /**
      * @param args the command line arguments
      */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(products.class
-                    .getName()).log(java.util.logging.Level.SEVERE, null, ex);
-
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(products.class
-                    .getName()).log(java.util.logging.Level.SEVERE, null, ex);
-
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(products.class
-                    .getName()).log(java.util.logging.Level.SEVERE, null, ex);
-
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(products.class
-                    .getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new products().setVisible(true);
-            }
-        });
-    }
+   
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JLabel accs;
